+++ sulong/CHANGELOG.md
@@ -28,6 +28,9 @@ Removed:
 * Removed the `--llvm.sourcePath` option (deprecated since 19.0), use
   `--inspect.SourcePath` instead.
 
+* Removed support for the application/x-llvm-ir-bitcode-base64 mime-type
+  (deprecated since 19.0).
+
 # Version 20.1.0
 
 Changes:
@@ -237,7 +240,7 @@ Changes:
 * Removed support for implicit polyglot types for local variables
   as the availability of type information is not guaranteed.
   Explicit polyglot casts are now strictly required (`polyglot_as_typed`).
-  See [docs/INTEROP.md](docs/INTEROP.md) and [polyglot.h](projects/com.oracle.truffle.llvm.libraries.bitcode/include/polyglot.h)
+  See [docs/contributor/INTEROP.md](docs/contributor/INTEROP.md) and [polyglot.h](projects/com.oracle.truffle.llvm.libraries.bitcode/include/polyglot.h)
   for more details.
 * Support for IR-level tracing.
 * Preliminary support for LLVM 7.